% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twitter_analytics_downloader.R
\name{twitter_analytics_downloader}
\alias{twitter_analytics_downloader}
\title{Tweet Activity}
\usage{
twitter_analytics_downloader(username, password, start_date, end_date,
  download_directory = getwd())
}
\arguments{
\item{username}{String specifying the phone, email or username of a Twitter account.}

\item{password}{String specifying the password of the Twitter account.}

\item{start_date}{String specifying the start date of the period to download in format "\%Y-\%m-\%d".}

\item{end_date}{String specifying the end date of the period to download in format "\%Y-\%m-\%d".}

\item{download_directory}{String specifying the directory where the downloaded files will be saved. Current directory is set by default.}
}
\description{
Downloads the Tweet activity report from Twitter Analytics.
}
\details{
This function downloads one or more csv files containing detailed data of the given
account's tweets: number of impressions, engagement rate, number of retweets,
link clicks, etc.

Twitter Analytics restricts the maximum possible date range to a 91-day window.
If the selected date range is longer than 91 days, this function splits the download
into several downloads.
}
\examples{
\dontrun{
twitter_analytics_downloader("username", "password", "2018-01-01", "2018-04-01", "C:/Users/ABCD/Downloads")}
}
